chore: proto visuel
