+++ src/client/components/PageRapportsHebdomadaires/PageRapportsHebdomadaires.styled.ts
@@ -1,0 +1,28 @@
+import styled from "@emotion/styled";
+
+const PageRapportsHebdomadairesStyled = styled.div`
+  background: var(--background-contrast-grey);
+
+  h2 {
+    color: var(--text-title-blue-france);
+  }
+
+  h3 {
+    color: var(--text-title-blue-france);
+  }
+
+  tbody {
+    background: none;
+  }
+
+  .horizontal-panel {
+    position: sticky;
+    top: 0;
+    z-index: 1;
+    width: 100%;
+    background-color: #f5f5fe;
+    box-shadow: 0 6px 18px var(--shadow-color);
+  }
+`;
+
+export default PageRapportsHebdomadairesStyled;
